initial support for extended nodes

+++ src/r_defs.h
@@ -284,7 +284,7 @@ typedef struct
 typedef struct subsector_s
 {
   sector_t *sector;
-  unsigned short numlines, firstline;
+  int numlines, firstline;
 } subsector_t;
 
 
@@ -295,7 +295,7 @@ typedef struct
 {
   fixed_t  x,  y, dx, dy;        // Partition line.
   fixed_t bbox[2][4];            // Bounding box for each child.
-  unsigned short children[2];    // If NF_SUBSECTOR its a subsector.
+  unsigned int children[2];      // If NF_SUBSECTOR its a subsector.
 } node_t;
 
 //
